+++ lib/braintree/validation_error.rb
@@ -6,8 +6,16 @@ module Braintree
     attr_reader :code
     attr_reader :message
 
-    def initialize(attributes)
-      set_instance_variables_from_hash attributes
+    def initialize(error_hash)
+      # parse GraphQL response objects
+      if (error_hash[:extensions] &&
+          error_hash[:extensions][:errorClass] &&
+          error_hash[:extensions][:errorClass] == "VALIDATION")
+        error_hash[:code] = error_hash[:extensions][:legacyCode].to_i
+        error_hash[:attribute] = error_hash[:path].last
+      end
+
+      set_instance_variables_from_hash error_hash
     end
 
     def inspect # :nodoc: