+++ lib/braintree/webhook_testing_gateway.rb
@@ -64,6 +64,8 @@ module Braintree
         _disbursement_sample_xml(id)
       when Braintree::WebhookNotification::Kind::SubscriptionChargedSuccessfully
         _subscription_charged_successfully(id)
+      when Braintree::WebhookNotification::Kind::SubscriptionChargedUnsuccessfully
+        _subscription_charged_unsuccessfully(id)
       when Braintree::WebhookNotification::Kind::AccountUpdaterDailyReport
         _account_updater_daily_report_sample_xml(id)
       when Braintree::WebhookNotification::Kind::ConnectedMerchantStatusTransitioned
@@ -76,6 +78,8 @@ module Braintree
         _ideal_payment_failed_sample_xml(id)
       when Braintree::WebhookNotification::Kind::GrantedPaymentInstrumentUpdate
         _granted_payment_instrument_update_sample_xml(id)
+      when Braintree::WebhookNotification::Kind::LocalPaymentCompleted
+        _local_payment_completed_sample_xml(id)
       else
         _subscription_sample_xml(id)
       end
@@ -95,7 +99,28 @@ module Braintree
           <id>#{id}</id>
           <transactions type="array">
             <transaction>
+              <id>#{id}</id>
               <status>submitted_for_settlement</status>
+              <amount>49.99</amount>
+            </transaction>
+          </transactions>
+          <add_ons type="array">
+          </add_ons>
+          <discounts type="array">
+          </discounts>
+        </subscription>
+      XML
+    end
+
+    def _subscription_charged_unsuccessfully(id)
+
+      <<-XML
+        <subscription>
+          <id>#{id}</id>
+          <transactions type="array">
+            <transaction>
+              <id>#{id}</id>
+              <status>failed</status>
               <amount>49.99</amount>
             </transaction>
           </transactions>
@@ -589,7 +614,6 @@ module Braintree
     end
 
     def _ideal_payment_complete_sample_xml(id)
-
       <<-XML
         <ideal-payment>
           <id>#{id}</id>
@@ -606,7 +630,6 @@ module Braintree
     end
 
     def _ideal_payment_failed_sample_xml(id)
-
       <<-XML
         <ideal-payment>
           <id>#{id}</id>
@@ -623,7 +646,6 @@ module Braintree
     end
 
     def _granted_payment_instrument_update_sample_xml(id)
-
       <<-XML
         <granted-payment-instrument-update>
           <grant-owner-merchant-id>vczo7jqrpwrsi2px</grant-owner-merchant-id>
@@ -639,6 +661,15 @@ module Braintree
             <item>expiration-year</item>
           </updated-fields>
         </granted-payment-instrument-update>
+      XML
+    end
+
+    def _local_payment_completed_sample_xml(id)
+      <<-XML
+        <local-payment>
+          <payment-id>PAY-XYZ123</payment-id>
+          <payer-id>ABCPAYER</payer-id>
+        </local-payment>
       XML
     end
   end